refactor(api): move config endpoints to players route
